Updated: moved extensions from controllers to module
